gcloud-launcher/pyproject.toml
[project]
name = "gcloud-launcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "google-cloud-aiplatform>=1.79.0",
    "python-dotenv>=1.0.1",
]
